+++ src/Error/ErrorInterface.php
@@ -16,8 +16,6 @@ interface ErrorInterface
      */
     public function getType(): ?string;
 
-    public function getStatusCode(): int;
-
     /**
      * @return array{class: non-empty-string, type?: non-empty-string}
      */
